binding generator is now a library
raylib_binding_generator it's the specific script that generates the
raylib binding.

Also, there is no need to manually tweak the binding, it's fully
functional (except, that is don't handle arrays, I still need to think
how to solve this problem).
