+++ binding_generator/binding_generator.lua
@@ -3,7 +3,6 @@
      file, You can obtain one at https://mozilla.org/MPL/2.0/. ]]
 
 local ins = require 'inspect'
-local lpeg_raylib_reader = require 'binding_generator/lpeg-raylib-reader'
 
 -- configure your generation here:
 local config = {
@@ -173,6 +172,7 @@ local function new_result(...)
    return result
 end
 
+-- table with functions to convert an AST node to nelua code string
 local converters = {}
 
 local function traverse(values, separator)
@@ -262,7 +262,6 @@ function converters.specifiers_and_qualifiers(value)
          list:remove(i)
       end
    end
-
 
    return new_result(list:concat())
 end
@@ -469,6 +468,7 @@ function converters.struct_declaration_list(value)
          if is_comment then
             list[i] = config.identation .. s .. '\n'
          elseif not space_only then
+            s = s:gsub('void%*', 'pointer')
             list[i] = config.identation .. s .. ','
          else
             list[i] = s
@@ -619,143 +619,7 @@ function converters.convert(subject)
    return result
 end
 
-local linklibs = {
-   'raylib',
-   'GL',
-   'glfw',
-   'openal',
-   'm',
-   'pthread',
-   'dl',
-   'X11',
-   'Xrandr',
-   'Xinerama',
-   'Xi',
-   'Xxf86vm',
-   'Xcursor'
+return {
+   converters = converters,
+   config = config
 }
-
-local cincludes = {
-   '<raylib.h>', '<raymath.h>'
-}
-
-local final_result = {}
-
-table.insert(final_result, '-- links: \n')
-
-for i = 1, #linklibs do
-   table.insert(final_result, "## linklib '" .. linklibs[i] .. "' \n")
-end
-
-table.insert(final_result, '\n-- includes: \n')
-
-for i = 1, #cincludes do
-   table.insert(final_result, "## cinclude '" .. cincludes[i] .. "' \n")
-end
-
--- TODO: Implement callback!
-table.insert(final_result, '\nglobal TraceLogCallback = @record{}\n')
-
--- [ raylib.h [
-config.record_in_use = 'Raylib'
-
-local raylib_table = lpeg_raylib_reader.read'binding_generator/modified-raylib.h'
-local raylib_result = converters.convert(raylib_table)
-
-table.insert(final_result, '\n-- raylib binding:\n')
-table.insert(final_result, 'global Raylib = @record{}')
-
-for i = 1, #raylib_result do
-   table.insert(final_result, raylib_result[i])
-end
-
-table.insert(final_result, '')
-
--- ] raylib.h ]
-
--- [ raymath.h [
-config.record_in_use = 'Raymath'
-
-local raymath_table = lpeg_raylib_reader.read'binding_generator/modified-raymath.h'
-local raymath_result = converters.convert(raymath_table)
-
-table.insert(final_result, '\n\n-- raymath binding: \n')
-table.insert(final_result, 'global Raymath = @record{}\n')
-
-for i = 1, #raymath_result do
-   table.insert(final_result, raymath_result[i])
-end
-
-table.insert(final_result, '\n')
-
--- TODO: split this script in different scripts and steps to not mix
--- "generic binding generation" and "raylib specific binding generation".
-table.insert(final_result, [[
--- [ operator overloading [
-
--- [ Vector2 [
--- Add two vectors (v1 + v2)
-function Vector2.__add(v1: Vector2, v2: Vector2): Vector2 <cimport'Vector2Add', nodecl> end
--- Subtract two vectors (v1 - v2)
-function Vector2.__sub(v1: Vector2, v2: Vector2): Vector2 <cimport'Vector2Subtract', nodecl> end
--- Calculate vector length
-function Vector2.__len(v: Vector2): float32 <cimport'Vector2Length', nodecl> end
--- Negate vector
-function Vector2.__unm(v: Vector2): Vector2 <cimport'Vector2Negate', nodecl> end
--- Divide vector by a float value or vector
-function Vector2.__div(v: Vector2, divisor: #[concept(function(d) return d.type.is_arithmetic or d.type.is_vector2 end)]#): Vector2
-  ## local fn_name = divisor.type.is_arithmetic and 'Divide' or 'DivideV'
-  return Vector2.#|fn_name|#(v, divisor)
-end
--- Scale vector (multiply by value) or Multiply vector by vector
-function Vector2.__mul(v: Vector2, multiplier: #[concept(function(m) return m.type.is_arithmetic or m.type.is_vector2 end)]#)
-  ## local fn_name = multiplier.type.is_arithmetic and 'Scale' or 'MultiplyV'
-  return Vector2.#|fn_name|#(v, multiplier)
-end
--- ] Vector2 ]
-
--- [ Vector3 [
--- Add two vectors
-function Vector3.__add(v1: Vector3, v2: Vector3): Vector3 <cimport'Vector3Add', nodecl> end
--- Subtract two vectors
-function Vector3.__sub(v1: Vector3, v2: Vector3): Vector3 <cimport'Vector3Subtract', nodecl> end
--- Calculate vector length
-function Vector3.__len(v: Vector3): float32 <cimport'Vector3Length', nodecl> end
--- Negate provided vector (invert direction)
-function Vector3.__unm(v: Vector3): Vector3 <cimport'Vector3Negate', nodecl> end
--- Multiply vector by scalar or by vector
-function Vector3.__mul(v: Vector3, multiplier: #[concept(function(m) return m.type.is_arithmetic or m.type.is_vector3 end)]#)
-  ## local fn_name = multiplier.type.is_arithmetic and 'Scale' or 'Multiply'
-  return Vector3.#|fn_name|#(v, multiplier)
-end
--- Divide vector by a float value or by vector
-function Vector3.__div(v: Vector3, divisor: #[concept(function(d) return d.type.is_arithmetic or d.type.is_vector3 end)]#)
-  ## local fn_name = divisor.type.is_arithmetic and 'Divide' or 'DivideV'
-  return Vector3.#|fn_name|#(v, divisor)
-end
--- ] Vector3 ]
-
--- [ Matrix [
--- Add two matrices
-function Matrix.__add(left: Matrix, right: Matrix): Matrix <cimport'MatrixAdd', nodecl> end
--- Subtract two matrices (left - right)
-function Matrix.__sub(left: Matrix, right: Matrix): Matrix <cimport'MatrixSubtract', nodecl> end
--- Returns two matrix multiplication
--- NOTE: When multiplying matrices... the order matters!
-function Matrix.__mul(left: Matrix, right: Matrix): Matrix <cimport'MatrixMultiply', nodecl> end
--- ] Matrix ]
-
--- [ Quaternion [
--- Computes the length of a quaternion
-function Quaternion.__len(q: Quaternion): float32 <cimport'QuaternionLength', nodecl> end
--- Calculate two quaternion multiplication
-function Quaternion.__mul(q1: Quaternion, q2: Quaternion): Quaternion <cimport'QuaternionMultiply', nodecl> end
--- ] Quaternion ]
-
--- ] operator overloading ]
-]])
---] raymath.h ]
-
-local file_to_generate = io.open('generated-raylib.nelua', 'w+')
-file_to_generate:write(table.concat(final_result))
-file_to_generate:close()
